See notes, - moved tests into logical files - removed large amount of duplication - shaved 2000ms of time to run - moved non tests into stubs folder - moved benchmark code into bench folder
