+++ test/bench/bench-external.js
