+++ test/bench/bench-internal.js
