+++ test/stubs/client-foo-pin.js
