+++ test/stubs/client-foo.js
