+++ test/stubs/fault.js
