+++ test/stubs/foo.js
