+++ test/stubs/readme-color-client.js
