+++ test/stubs/readme-color-service.js
