+++ test/stubs/readme-color-tcp.js
