+++ test/stubs/readme-color-web-https.js
