+++ test/stubs/readme-color.js
