+++ test/stubs/readme-many-colors-client.js
