+++ test/stubs/readme-many-colors-server.js
