+++ test/stubs/readme-many-colors.sh
