+++ test/stubs/service-foo.js
